feat: Criado serviço para notificações
